jira-spark-etl-master/pyproject.toml
[tool.poetry]
name = "jira_database_etl"
version = "0.1.0"
description = "Script to import issues from a JIRA instance into a database."
authors = ["ashish_jha@optum.com>"]
maintainers = ["Ashish Jha <ashish_jha@optum.com>>"]
readme = "README.md"
repository = "https://github.com/ajha37/jira-database-etl/"
keywords = [
    "Jira",
    "ETL",
    "Database",
    "Python",
    "Tutorial"
]

[tool.poetry.dependencies]
python = "^3.8"
requests = "*"
pymysql = "*"
pandas = "*"
python-dotenv = "*"
loguru = "*"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
run = "main:init_script"

[tool.poetry.urls]
issues = "https://github.com/ajha37/jira-database-etl/issues"